+++ cli/src/test/java/com/box/l10n/mojito/cli/checks/PlaceholderCommentCheckerTest.java
@@ -1,4 +1,4 @@
-package com.box.l10n.mojito.cli.checker;
+package com.box.l10n.mojito.cli.checks;
 
 import com.box.l10n.mojito.cli.command.checks.CliCheckResult;
 import com.box.l10n.mojito.cli.command.checks.CliCheckerOptions;
